requirements copy.txt
# Создаем requirements.txt с основными пакетами
echo pandas==2.0.3 > requirements.txt
echo matplotlib==3.7.2 >> requirements.txt
echo numpy==1.24.3 >> requirements.txt
echo requests==2.31.0 >> requirements.txt
echo seaborn==0.12.2 >> requirements.txt
echo scipy==1.10.1 >> requirements.txt
echo statsmodels==0.14.0 >> requirements.txt
echo tinkoff-invest==1.0.12 >> requirements.txt
python-dotenv>=1.0.0
pandas>=1.5.0
psutil>=5.9.0
tinkoff-investments>=0.6.0

